requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==0.19.2

